Implement a more comprehensive encoder and aux port system
Each of the two auxiliary ports can now have its pins individually
configured and assigned to many different options for inputs,
including SPI, UART, quadrature, sine/cosine, and hall effects.

The auxiliary port inputs can be configured individually to produce
theta values for commutation, and output position/velocity for
position mode control.  Additionally, flexible options are provided to
use index inputs either for commutation, or as a homing switch on the
output.

Position mode control is moved to operate purely in a local relative
coordinate system.  This means that rezeroing and forcing the output
position can now occur without affecting control.  Only the next
command will take into account the new values (aside from optional
absolute position limits).

Position mode control has a new optional per-command override to enter
"fixed voltage" mode, which can be useful to perform homing on
incremental commutation sources.

Each encoder source has a 32 position compensation table that can be
used to adjust for non-linearities.  Preliminary scripts are present
to generate tables, but they are not yet integrated into moteus_tool.

Torque production has a 1024 position compensation table indexed to
the commutation encoder which can be used to implement anti-cogging
support.  Preliminary scripts are present to generate tables, but they
are not yet integrated into moteus_tool.

Each of the auxiliary port pins, when not used for encoders, may be
configured as GPIO, or for some, as ADC inputs.  These may be accessed
over the diagnostic or register protocol.

The maximum control loop rate has been lowered from 40kHz to 30kHz,
because, well, we're doing a lot more per control loop cycle.  The PWM
rate range remains unchanged at 15kHz to 60kHz.

Two new diagnostic mode commands "d hstart" "d hend" can be used to
record onboard histograms at millisecond intervals with configurable x
and y axes.  This can be used to implement various calibration
techniques even with a slow host.

This revises the ELF version to 0x0105 because of the new
configuration, and the register protocol version to 5, because of the
changed semantics of the "rezero state" register.

moteus_tool attempts to port configuration from old versions to the
new.  Its backporting ability is largely limited to those
configurations which it generated itself as a result of an upgrade.

The complete list of external encoders supported as of this commit:
 * SPI
   * Onboard AS5047
   * External AS5047
   * iC-PZ
 * I2C
   * AS5048
   * AS5600
 * Quadrature
 * Sine/Cosine
 * Hall effect
 * UART
   * RLS AksIM-2